more cases of WordPress.Security.ValidatedSanitizedInput.InputNotSanitized

+++ lib/admin/groups-admin-groups.php
@@ -68,7 +68,7 @@ function groups_admin_groups() {
 				break;
 			case 'edit' :
 				if ( !( $group_id = groups_admin_groups_edit_submit() ) ) {
-					return groups_admin_groups_edit( $_POST['group-id-field'] );
+					return groups_admin_groups_edit( $_POST['group-id-field'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 				} else {
 					$group = Groups_Group::read( $group_id );
 					Groups_Admin::add_message( sprintf(
@@ -84,26 +84,26 @@ function groups_admin_groups() {
 				break;
 			// bulk actions on groups: add capabilities, remove capabilities, remove groups
 			case 'groups-action' :
-				if ( wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_ACTION_NONCE], 'admin' ) ) {
-					$group_ids = isset( $_POST['group_ids'] ) ? $_POST['group_ids'] : null;
+				if ( wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_ACTION_NONCE], 'admin' ) ) { // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+					$group_ids = isset( $_POST['group_ids'] ) ? $_POST['group_ids'] : null; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 					$bulk_action = null;
 					if ( isset( $_POST['bulk'] ) ) {
-						$bulk_action = $_POST['bulk-action'];
+						$bulk_action = $_POST['bulk-action']; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 					}
 					if ( is_array( $group_ids ) && ( $bulk_action !== null ) ) {
 						foreach ( $group_ids as $group_id ) {
 							switch ( $bulk_action ) {
 								case 'add-capability' :
-									$capabilities_id = isset( $_POST['capability_id'] ) ? $_POST['capability_id'] : null;
-									if ( $capabilities_id !== null ) {
+									$capabilities_id = isset( $_POST['capability_id'] ) ? $_POST['capability_id'] : null; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+									if ( $capabilities_id !== null && is_array( $_POST['capability_id'] ) ) {
 										foreach ( $capabilities_id as $capability_id ) {
 											Groups_Group_Capability::create( array( 'group_id' => $group_id, 'capability_id' => $capability_id ) );
 										}
 									}
 									break;
 								case 'remove-capability' :
-									$capabilities_id = isset( $_POST['capability_id'] ) ? $_POST['capability_id'] : null;
-									if ( $capabilities_id !== null ) {
+									$capabilities_id = isset( $_POST['capability_id'] ) ? $_POST['capability_id'] : null; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+									if ( $capabilities_id !== null && is_array( $_POST['capability_id'] ) ) {
 										foreach ( $capabilities_id as $capability_id ) {
 											Groups_Group_Capability::delete( $group_id, $capability_id );
 										}
@@ -131,12 +131,12 @@ function groups_admin_groups() {
 				break;
 			case 'edit' :
 				if ( isset( $_GET['group_id'] ) ) {
-					return groups_admin_groups_edit( $_GET['group_id'] );
+					return groups_admin_groups_edit( $_GET['group_id'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 				}
 				break;
 			case 'remove' :
 				if ( isset( $_GET['group_id'] ) ) {
-					return groups_admin_groups_remove( $_GET['group_id'] );
+					return groups_admin_groups_remove( $_GET['group_id'] ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 				}
 				break;
 		}
@@ -150,7 +150,7 @@ function groups_admin_groups() {
 		isset( $_POST['group_id'] ) ||
 		isset( $_POST['group_name'] )
 	) {
-		if ( !wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_FILTER_NONCE], 'admin' ) ) {
+		if ( !wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_FILTER_NONCE], 'admin' ) ) { // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 			wp_die( esc_html__( 'Access denied.', 'groups' ) );
 		}
 	}
@@ -167,7 +167,7 @@ function groups_admin_groups() {
 	} else if ( isset( $_POST['submitted'] ) ) {
 		// filter by name
 		if ( !empty( $_POST['group_name'] ) ) {
-			$group_name = $_POST['group_name'];
+			$group_name = sanitize_text_field( $_POST['group_name'] );
 			Groups_Options::update_user_option( 'groups_group_name', $group_name );
 		}
 		// filter by group id
@@ -181,18 +181,18 @@ function groups_admin_groups() {
 	}
 
 	if ( isset( $_POST['row_count'] ) ) {
-		if ( !wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_NONCE_1], 'admin' ) ) {
+		if ( !wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_NONCE_1], 'admin' ) ) { // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 			wp_die( esc_html__( 'Access denied.', 'groups' ) );
 		}
 	}
 
 	if ( isset( $_POST['paged'] ) ) {
-		if ( !wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_NONCE_2], 'admin' ) ) {
+		if ( !wp_verify_nonce( $_POST[GROUPS_ADMIN_GROUPS_NONCE_2], 'admin' ) ) { // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 			wp_die( esc_html__( 'Access denied.', 'groups' ) );
 		}
 	}
 
-	$current_url = ( is_ssl() ? 'https://' : 'http://' ) . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
+	$current_url = ( is_ssl() ? 'https://' : 'http://' ) . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI']; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 	$current_url = remove_query_arg( 'paged', $current_url );
 	$current_url = remove_query_arg( 'action', $current_url );
 	$current_url = remove_query_arg( 'group_id', $current_url );
@@ -238,7 +238,7 @@ function groups_admin_groups() {
 		$paged = 0;
 	}
 
-	$orderby = isset( $_GET['orderby'] ) ? $_GET['orderby'] : null;
+	$orderby = isset( $_GET['orderby'] ) ? $_GET['orderby'] : null; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 	switch ( $orderby ) {
 		case 'group_id' :
 		case 'name' :
@@ -248,7 +248,7 @@ function groups_admin_groups() {
 			$orderby = 'name';
 	}
 
-	$order = isset( $_GET['order'] ) ? $_GET['order'] : null;
+	$order = isset( $_GET['order'] ) ? $_GET['order'] : null; // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 	switch ( $order ) {
 		case 'asc' :
 		case 'ASC' :
